requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.14
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
frozenlist==1.7.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
makefun==1.16.0
mako==1.3.10
markupsafe==3.0.2
multidict==6.6.2
packaging==25.0
propcache==0.3.2
psycopg2-binary==2.9.10
pwdlib==0.2.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyjwt==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typing-extensions==4.14.0
typing-inspection==0.4.1
uvicorn==0.35.0
yarl==1.20.1
